app/requirements.txt
Flask
paho-mqtt==1.5.1
# Ajoutez d'autres dépendances si nécessaire
